requirements.txt
discord.py[voice]>=2.7.1
gTTS==2.5.1
python-dotenv==1.0.1
